requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
boto3==1.26.114
botocore==1.29.114
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.5
django-admin-logs==1.2.0
django-admin-rangefilter==0.9.0
django-cors-headers==3.14.0
django-crontab==0.7.1
django-debug-toolbar==4.1.0
django-dotenv==1.4.2
django-environ==0.9.0
django-import-export==3.0.2
django-ratelimit==4.1.0
django-redis==5.2.0
django-rest-swagger==2.2.0
django-ses==3.5.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
ecs-logging==2.1.0
elastic-apm==6.20.0
et-xmlfile==1.1.0
Faker==18.3.1
guacamole-api-wrapper==0.0.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupPy==1.14
MarkupSafe==2.1.2
mysqlclient==2.1.1
numpy==1.24.2
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.1.0
packaging==23.1
pandas==1.5.3
Pillow==9.4.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
redis==4.5.4
requests==2.30.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
sentry-sdk==1.18.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.3
tablib==3.3.0
uritemplate==4.1.1
urllib3==1.26.15
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
